refactor: seperate state and component
