+++ simaple/simulate/component/specific/magician.py
@@ -2,11 +2,8 @@ from typing import Optional
 
 from simaple.core.base import Stat
 from simaple.simulate.component.base import reducer_method, view_method
-from simaple.simulate.component.skill import (
-    CooldownState,
-    DurationState,
-    SkillComponent,
-)
+from simaple.simulate.component.skill import SkillComponent
+from simaple.simulate.component.state import CooldownState, DurationState
 from simaple.simulate.component.view import Running, Validity
 from simaple.simulate.global_property import Dynamics
